added an if statement in the migration file

+++ db/migrate/20241101225314_remove_band_id_from_chats.rb
@@ -1,5 +1,7 @@
 class RemoveBandIdFromChats < ActiveRecord::Migration[7.1]
   def change
-    remove_column :chats, :band_id, :bigint
+    if column_exists?(:chats, :band_id)
+      remove_column :chats, :band_id, :bigint
+    end
   end
 end
